New paginator test + some bug fixes and refactoring + cleanup
