+++ dataland-frontend/tests/e2e/specs/low/pages/taxonomy/IndexPanelTabMenu.ts
@@ -1,13 +1,22 @@
 import apiSpecs from "@/../build/clients/backend/backendOpenApi.json";
+import Chainable = Cypress.Chainable;
+
 const numberOfStockIndices = apiSpecs.components.schemas.CompanyInformation.properties["indices"].items.enum.length
+
+export function checkIfDaxTabIsHighlighted(): Chainable<JQuery> {
+    return cy.get('li[class="p-tabmenuitem p-highlight"]')
+        .children('.p-menuitem-link')
+        .children('.p-menuitem-text')
+        .should('contain', 'DAX')
+}
 
 describe('Index Panel behavior', function () {
     const indexTabMenu = '.p-tabmenu > .p-tabmenu-nav > .p-tabmenuitem > .p-menuitem-link'
-    beforeEach(()=> {
+    beforeEach(() => {
         cy.restoreLoginSession()
     })
 
-    it('Index tabmenu should be present on first visit', () => {
+    it('Index tabmenu should be present', () => {
         cy.visit("/searchtaxonomy")
         cy.get('.p-tabmenuitem').should('have.length', numberOfStockIndices)
         cy.get(indexTabMenu)
@@ -17,10 +26,10 @@ describe('Index Panel behavior', function () {
         cy.get('h2')
             .should('contain', "Results")
         cy.get('table.p-datatable-table').should('exist')
-        cy.get('table.p-datatable-table').contains('th','COMPANY')
-        cy.get('table.p-datatable-table').contains('th','SECTOR')
-        cy.get('table.p-datatable-table').contains('th','MARKET CAP')
-        cy.get('table.p-datatable-table').contains('td','VIEW')
+        cy.get('table.p-datatable-table').contains('th', 'COMPANY')
+        cy.get('table.p-datatable-table').contains('th', 'SECTOR')
+        cy.get('table.p-datatable-table').contains('th', 'MARKET CAP')
+        cy.get('table.p-datatable-table').contains('td', 'VIEW')
             .contains('a', 'VIEW')
             .click()
             .url().should('include', '/companies/')
@@ -29,24 +38,15 @@ describe('Index Panel behavior', function () {
     });
 
 
+    it('Visit searchtaxonomy page, scroll to the bottom, back to the top, and check if Dax still highlighted', () => {
+        cy.restoreLoginSession()
+        cy.visit('/searchtaxonomy')
 
-    it('Index tabmenu should be present', () => {
-        cy.visit("/searchtaxonomy")
-        cy.get(indexTabMenu).should('exist')
-        cy.get('.grid')
-            .should('not.contain','Choose by stock market index')
-        cy.get('h2')
-            .should('contain', "Results")
-        cy.get('table.p-datatable-table').should('exist')
-        cy.get('table.p-datatable-table').contains('th','COMPANY')
-        cy.get('table.p-datatable-table').contains('th','SECTOR')
-        cy.get('table.p-datatable-table').contains('th','MARKET CAP')
-        cy.get('table.p-datatable-table').contains('td','VIEW')
-            .contains('a', 'VIEW')
-            .click()
-            .url().should('include', '/companies/')
-            .url().should('include', '/eutaxonomies')
+        checkIfDaxTabIsHighlighted()
+
+        cy.scrollTo('bottom', {duration: 500})
+        cy.scrollTo('top', {duration: 500})
+
+        checkIfDaxTabIsHighlighted()
     });
-
-
 });
