+++ dataland-frontend/tests/e2e/specs/low/pages/taxonomy/Paginator.ts
@@ -1,0 +1,24 @@
+import Chainable = Cypress.Chainable;
+
+describe('Paginator test suite', () => {
+
+    it('Assure existence of paginator for Dax search', () => {
+        cy.restoreLoginSession()
+        cy.visit('/searchtaxonomy')
+        cy.get('#app').should("exist")
+        cy.get('h1').should("contain", "Search EU Taxonomy data")
+        cy.get('div.p-paginator').should('exist')
+    });
+
+
+    it('Do a search with 0 matches, then assure that the paginator is gone', () => {
+        cy.visit("/searchtaxonomy")
+        const inputValue = "ABCDEFGHIJKLMNOPQRSTUVWXYZ12345678987654321"
+        cy.get('input[name=eu_taxonomy_search_input]')
+            .should('exist')
+            .type(inputValue)
+            .type('{enter}')
+            .should('have.value', inputValue)
+        cy.get('div.p-paginator').should('not.exist')
+    });
+})
